Extract AddSubmenu() into it’s own module
