+++ cmd2/__init__.py
@@ -1,4 +1,4 @@
 #
 # -*- coding: utf-8 -*-
-from .cmd2 import __version__, Cmd, AddSubmenu, CmdResult, categorize
+from .cmd2 import __version__, Cmd, CmdResult, categorize
 from .cmd2 import with_argument_list, with_argparser, with_argparser_and_unknown_args, with_category
